Split Functions into Modules
* Start to split functions into separate modules
* Fixed an issue in the main script where it loads the old fine tune model
